fix: remove duplicate import
